requirements.txt
mcp>=1.6.0
httpx>=0.27.0
beautifulsoup4>=4.12.3
python-dotenv>=1.0.1
pypdf>=5.0.0
lxml>=5.2.0
playwright>=1.52.0
